moved classes to their own files on lib folder
