+++ lib/player.rb
@@ -1,0 +1,7 @@
+class Player
+  attr_accessor :name, :taken
+  def initialize(name = 'Player')
+    @name = name
+    @taken = []
+  end
+end
